+++ examples/vue3-vite/vite.config.ts
@@ -6,7 +6,6 @@ export default defineConfig({
   plugins: [
     vue(),
     fileViewerRenderers({
-      preset: 'all',
       copyAssets: true
     })
   ]